+++ internal/plugins/db/fsdb/env_lock_windows.go
@@ -1,0 +1,19 @@
+//go:build windows
+
+package fsdb
+
+import (
+	"os"
+
+	"golang.org/x/sys/windows"
+)
+
+func lockExclusive(f *os.File) error {
+	var overlapped windows.Overlapped
+	return windows.LockFileEx(windows.Handle(f.Fd()), windows.LOCKFILE_EXCLUSIVE_LOCK, 0, 1, 0, &overlapped)
+}
+
+func unlockExclusive(f *os.File) error {
+	var overlapped windows.Overlapped
+	return windows.UnlockFileEx(windows.Handle(f.Fd()), 0, 1, 0, &overlapped)
+}